[Refactor] Directory structur e
