+++ index.html
@@ -9,8 +9,8 @@
 <body>
     <!-- htmlのcanvas要素を使用 -->
     <!-- <canvas id="canvas" width="300" height="660"></canvas> -->
-    <canvas id="canvas" ></canvas>
-    <script src="js/main.js"></script>
+    <canvas id="canvas"></canvas>
+    <script type="module" src="js/main.js"></script>
 </body>
 
 </html>
