+++ js/board.js
@@ -1,0 +1,3 @@
+import { PLAYSCREENWIDTH, PLAYSCREENHEIGHT } from './utils.js';
+
+// ゲームボードの状態を管理するロジックをここに実装
