+++ js/utils.js
@@ -1,0 +1,10 @@
+// 1ブロックの大きさ
+export const BLOCKSIZE = 30;
+// フィールドのサイズ
+export const PLAYSCREENWIDTH = 10;
+export const PLAYSCREENHEIGHT = 22;
+// キャンバスサイズ（＝プレイ画面のサイズ）
+export const CANVASWIDTH = BLOCKSIZE * PLAYSCREENWIDTH;
+export const CANVASHEIGHT = BLOCKSIZE * PLAYSCREENHEIGHT;
+
+// 他の汎用的な関数や定数をここに定義
